User KEY TYPE Integer로 변경
User엔티티 seq 타입이 Integer임

+++ demo/src/main/java/com/example/demo/repository/UserInterface.java
@@ -16,7 +16,7 @@ import java.util.Optional;
 import java.util.Set;
 
 @Repository
-public interface UserInterface extends JpaRepository<User, Long> {
+public interface UserInterface extends JpaRepository<User, Integer> {
     Long countByCreateAtGreaterThanEqual(Date currentDate);
     Long countByBlockYn(String blockYn);
     List<User> findByName(String name);
